requirements.txt
black==25.11.0
click==8.3.0
colorama==0.4.6
flake8==7.3.0
iniconfig==2.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.1
pytokens==0.3.0
PyYAML==6.0.3
